Organize Pokemon components into logical folder structure
- Create pokemon/list/ and pokemon/detail/ directories for better component organization
- Move 6 list-related components to list/ directory (PokemonGrid, PokemonCard, etc.)
- Move 16 detail-related components to detail/ directory (PokemonBasicInfo, PokemonMoves, etc.)
- Update all import paths in main application files and cross-directory references
- Fix TypeScript import resolution for common UI components (Badge, DataEmptyState, etc.)
- Maintain sprites/ and evolution/ subdirectories for specialized component types
- Improve code maintainability through logical separation of list vs detail page components
