+++ client/components/ui/pokemon/detail/PokemonBasicInfo.tsx
