+++ client/components/ui/pokemon/detail/PokemonBasicInfoGrid.tsx
