+++ client/components/ui/pokemon/detail/PokemonDetailHeader.tsx
