+++ client/components/ui/pokemon/detail/PokemonDetailTabs.tsx
