+++ client/components/ui/pokemon/detail/PokemonHeader.tsx
