+++ client/components/ui/pokemon/detail/PokemonImage.tsx
