+++ client/components/ui/pokemon/detail/PokemonImageDisplay.tsx
