+++ client/components/ui/pokemon/detail/PokemonNavigation.tsx
