+++ client/components/ui/pokemon/detail/PokemonStatBar.tsx
