+++ client/components/ui/pokemon/detail/PokemonStatsSection.tsx
