+++ client/components/ui/pokemon/detail/PokemonStorySection.tsx
