+++ client/components/ui/pokemon/detail/PokemonWeaknessSection.tsx
